tests/wpt/tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-play.html

<!doctype html>
<title>invoking resource selection with play()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var v;
var t = async_test(function(t) {
  v = document.createElement('video');
  assert_equals(v.networkState, v.NETWORK_EMPTY, 'networkState after creating v');
  v.play();
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState after v.play()');
});
</script>
<script>
t.step(function() {
  assert_equals(v.networkState, v.NETWORK_EMPTY, 'networkState in separate script');
  t.done();
});
</script>
